Refactor / consolidate
