+++ app/pages/SiloAccessGroupsTab.tsx
@@ -33,6 +33,7 @@ import { titleCrumb } from '~/hooks/use-crumbs'
 import { confirmDelete } from '~/stores/confirm-delete'
 import { EmptyCell } from '~/table/cells/EmptyCell'
 import { ButtonCell } from '~/table/cells/LinkCell'
+import { MemberCountCell } from '~/table/cells/MemberCountCell'
 import { useColsWithActions, type MenuAction } from '~/table/columns/action-col'
 import { Columns } from '~/table/columns/common'
 import { useQueryTable } from '~/table/QueryTable'
@@ -57,11 +58,6 @@ export async function clientLoader() {
 export const handle = titleCrumb('Groups')
 
 const colHelper = createColumnHelper<Group>()
-
-function MemberCountCell({ groupId }: { groupId: string }) {
-  const { data } = useQuery(q(api.userList, { query: { group: groupId, limit: ALL_ISH } }))
-  return data ? <>{data.items.length}</> : null
-}
 
 const GroupEmptyState = () => (
   <EmptyMessage
